requirements.txt
sqlalchemy
gino
asyncpg
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv
